Update game.js
Here I canged the acceleration just a little bit and I changed for the robot to move up when pressing the space bar rather than the up button.

I also made the robot stop when reaching the ground.
________________________________

I then added a function startscreen, gamescreen (moving the game background to the game screen) and I made a resultscreen.

I put the some if states within the draw function for the different screens. Then is the mouse is clicked then the game will go from start screen to game and from game to result screen.